@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sample.AspNetCore.SystemTests.Test.Helpers
+﻿namespace Sample.AspNetCore.SystemTests.Test.Helpers
 {
     public class Checkout
     {
