@@ -4,6 +4,7 @@ using Atata;
 using NUnit.Framework;
 using Sample.AspNetCore.SystemTests.Services;
 using Sample.AspNetCore.SystemTests.Test.Helpers;
+using SwedbankPay.Sdk;
 
 namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Abort
 {
@@ -16,6 +17,7 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Abort
 
 
         [Test]
+        [Retry(3)]
         [TestCaseSource(nameof(TestData), new object[] { true, null })]
         public async Task Abort_PaymentOrder(Product[] products)
         {
@@ -29,7 +31,7 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Abort
             var order = await SwedbankPayClient.PaymentOrders.Get(new Uri(orderLink, UriKind.RelativeOrAbsolute), SwedbankPay.Sdk.PaymentOrders.PaymentOrderExpand.All);
 
             // Operations
-            Assert.That(order.Operations, Is.Empty);
+            Assert.That(order.Operations[LinkRelation.AbortedPaymentOrder], Is.Not.Null);
 
             // Transactions
             Assert.That(order.PaymentOrderResponse.CurrentPayment.Payment, Is.Null);
