@@ -3,7 +3,6 @@ using NUnit.Framework;
 using Sample.AspNetCore.SystemTests.Services;
 using Sample.AspNetCore.SystemTests.Test.Helpers;
 using SwedbankPay.Sdk;
-using SwedbankPay.Sdk.Payments;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -18,6 +17,7 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Anonymous
 
 
         [Test]
+        [Retry(3)]
         [TestCaseSource(nameof(TestData), new object[] { false, PaymentMethods.Swish })]
         public async Task Anonymous_PaymentOrder_Swish_Sale(Product[] products, PayexInfo payexInfo)
         {
