@@ -1,14 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using SwedbankPay.Sdk;
-using SwedbankPay.Sdk.Payments.CardPayments;
-using SwedbankPay.Sdk.Payments.SwishPayments;
-
-namespace Sample.AspNetCore.Models.ViewModels
+﻿namespace Sample.AspNetCore.Models.ViewModels
 {
     public class PaymentMenuViewModel
     {
