@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Primitives;
-using Moq;
+using NSubstitute;
 
 namespace SwedbankPay.Sdk.Tests.TestHelpers
 {
@@ -30,20 +30,20 @@ namespace SwedbankPay.Sdk.Tests.TestHelpers
 
         public IDirectoryContents GetDirectoryContents(string subpath)
         {
-            var contents = new Mock<IDirectoryContents>();
-            contents.Setup(m => m.Exists).Returns(true);
+            var contents = Substitute.For<IDirectoryContents>();
+            contents.Exists.Returns(true);
 
             if (string.IsNullOrEmpty(subpath))
             {
-                contents.Setup(m => m.GetEnumerator()).Returns(this.files.GetEnumerator());
-                return contents.Object;
+                contents.GetEnumerator().Returns(this.files.GetEnumerator());
+                return contents;
             }
 
             var filesInFolder = this.files.Where(f => f.Name.StartsWith(subpath, StringComparison.Ordinal));
             if (filesInFolder.Any())
             {
-                contents.Setup(m => m.GetEnumerator()).Returns(filesInFolder.GetEnumerator());
-                return contents.Object;
+                contents.GetEnumerator().Returns(filesInFolder.GetEnumerator());
+                return contents;
             }
             return NotFoundDirectoryContents.Singleton;
         }
