@@ -10,26 +10,12 @@ namespace SwedbankPay.Sdk.Tests.TestHelpers
 
         public static IConfigurationRoot GetIConfigurationRoot(string outputPath)
         {
-            #if DEBUG
             return new ConfigurationBuilder()
                 .SetBasePath(outputPath)
                 .AddJsonFile("appsettings.json", true)
                 .AddUserSecrets("55739ea0-5447-45e4-b35e-e0412f172f5f")
+                .AddEnvironmentVariables()
                 .Build();
-
-            #elif RELEASE
-
-            var json = $@"{{ ""SwedbankPayConnectionSettings"": {{ ""Token"": ""{Environment.GetEnvironmentVariable("SwedbankPayConnectionSettings.Token", EnvironmentVariableTarget.User)}""}}}}";
-            var memoryJsonFile = new MemoryFileInfo("config.json", System.Text.Encoding.UTF8.GetBytes(json), DateTimeOffset.Now);
-            var memoryFileProvider = new MockFileProvider(memoryJsonFile);
-
-            return new ConfigurationBuilder()
-                .SetBasePath(outputPath)
-                .AddJsonFile("appsettings.json", true)
-                .AddJsonFile(memoryFileProvider, "config.json", false, false)
-                .Build();
-
-            #endif
         }
 
 
@@ -41,7 +27,7 @@ namespace SwedbankPay.Sdk.Tests.TestHelpers
             var iConfig = GetIConfigurationRoot(outputPath);
 
             iConfig
-                .GetSection("SwedbankPayConnectionSettings")
+                .GetSection("SwedbankPay")
                 .Bind(configuration);
 
             return configuration;
